Use width and height PPI

+++ src/math.rs
@@ -1,4 +1,4 @@
-pub const MM_PER_INCH: f32 = 25.4;
+pub const MM_PER_INCH: f64 = 25.4;
 
 #[derive(Default, PartialEq)]
 pub struct Point {
